refactor: move sorted array methods to foundation
