+++ yarn-project/foundation/src/array/index.ts
@@ -1,1 +1,2 @@
 export * from './array.js';
+export * from './sorted_array.js';
